requirements.txt
Django>=4.2,<5.0
python-telegram-bot>=20.0
